+++ cmd/file.go
@@ -1,17 +1,8 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-*/
 package cmd
 
 import (
 	"github.com/spf13/cobra"
 )
-
-//
-//
-// Define the CLI data for the file sub-command
-//
-//
 
 func newAddFileCmd() *cobra.Command {
 	addFileCmd := &cobra.Command{