+++ CHANGELOG.md
@@ -103,6 +103,11 @@
   validated, which allows for working with incomplete or even invalid files in a pipeline.
   The functionality is imported from the [go-apiops library](https://github.com/Kong/go-apiops).
   [#939](https://github.com/Kong/deck/pull/939)
+- Added a new command `file render` to render a final decK file. This will result in a file representing
+  the state as it would be synced online.
+  [#963](https://github.com/Kong/deck/pull/963)
+- Added a new flag `--format` to `file convert` to enable JSON output.
+  [#963](https://github.com/Kong/deck/pull/963)
 
 ### Fixes